pyproject.toml
[project]
name = "ct-cex-research"
version = "0.1.0"
description = "CT-GSSN & CEX-TSLM reference implementations"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [{name = "Unaffiliated Research Group"}]

dependencies = [
  "torch>=2.1.0",
  "numpy>=1.24.0",
  "pandas>=1.5.0",
  "scikit-learn>=1.1.0",
  "plotly>=5.18.0",
  "tensorboard>=2.11.0",
  "pyyaml>=6.0",
  "tqdm>=4.65.0",
]

[project.optional-dependencies]
all = [
  "transformers>=4.40.0",
  "datasets>=2.16.0",
  "wandb>=0.16.0",
]

[tool.setuptools]
packages = ["ct_gssn", "cex_tslm"]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["tests"]
